Implement SetBufferLayout, use system thread for endpoint initialization
This actually breaks endpoint creation right now because the CLSID set
is returning STATUS_ACCESS_DENIED, but implement SetBufferLayout to set
buffer parameters and move the IoSetDeviceInterfacePropertyData calls to
a system thread per the documentation. Also move device enablement to
the KS minidriver PostStart callback instead of doing it in AddDevice.