Implement support for non-frame-aligned buffer sizes.
audiokse.dll does not seem to like the buffer rounding that was being
done before. It would sometimes fail with the undocumented HRESULT code
887C0057 on some buffer sizes some of the time. mux/demux can now handle
wrapping around the end of the ring buffer.

Also fixes a bug in SAR_START_REGISTRY_FILTER that would cause a bugcheck
the second time it was called.