Issue #13: debug logging, work on generalized registry filter
Adds some temporary debug logging to SarMMDeviceEnumerator to see
more API calls that clients are making. Since the problem with XAudio2
apps is its internal use of ActivateAudioInterfaceAsync, generalize the
registry filter to store mapped registry paths in a table instead of
just statically comparing strings.

Stub out an interceptor for ActiveAudioInterfaceWorker coclass.

All of this is pretty untested code.