Driver IOCTL: Use consistent KS/SAR dispatch
In IRP_MJ_CREATE, we are checking the target filename against the reference
filename used to control SAR.
Use the same check for other IRP instead of trial and error.
Sometimes, the order of IRP_MJ_CLEANUP/CLOSE between SAR control and KS
endpoints is not the expected one and could lead to unexpected endpoint
and controlContext release.

Like the controlContext not being released and leaving filters enabled
in the KSDevice. In that case, subsequent endpoint creation would not work
because KS still use the old factory (which was not in the endpoint list).