Driver: fixup some things
This may not have any incidence, but just in case.
Use KsFilterFactoryUpdateCacheData to refresh any possible caching (this
does not fix the cache behavior of windows 10).

Check only one time the expected buffer size and return STATUS_BUFFER_OVERFLOW
if it is 0.
Fill irp->IoStatus.Information when returning STATUS_BUFFER_OVERFLOW.

Use ProbeForRead in SarReadUserBuffer only when the caller is not the kernel
as the buffer won't be in user memory.